# docs/swagger.yaml
basePath: /
definitions:
  handler.CreateWallet:
    properties:
      user_id:
        type: integer
    required:
    - user_id
    type: object
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
  handler.getAllWalletsAdminResponce:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Wallet'
        type: array
    type: object
  handler.getAllWalletsResponce:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Wallet'
        type: array
    type: object
  handler.signInInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.BlockWallet:
    properties:
      block:
        type: boolean
      walletUUID:
        type: string
    required:
    - block
    - walletUUID
    type: object
  models.User:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.Wallet:
    properties:
      balance:
        type: integer
      blocked:
        type: boolean
      uuid:
        type: string
    type: object
  models.WalletUpdate:
    properties:
      amount:
        type: integer
      operationType:
        description: DEPOSIT or WITHDRAW
        type: string
      walletUUID:
        type: string
    required:
    - amount
    - operationType
    - walletUUID
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Server for Wallet
  title: Wallet API
  version: "1.0"
paths:
  /api/v1/admin/wallet:
    post:
      consumes:
      - application/json
      description: update wallet
      operationId: update-wallet-admin
      parameters:
      - description: Wallet query
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.WalletUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Wallet
      tags:
      - ADMIN
  /api/v1/admin/wallet/{user_id}:
    get:
      consumes:
      - application/json
      description: get all wallets by user id
      operationId: get-all-wallets-admin
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllWalletsAdminResponce'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get All Wallets By User ID
      tags:
      - ADMIN
  /api/v1/admin/wallet/block:
    put:
      consumes:
      - application/json
      description: block wallet
      operationId: block-wallet-admin
      parameters:
      - description: Block Wallet
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.BlockWallet'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Block Wallet
      tags:
      - ADMIN
  /api/v1/admin/wallet/new:
    post:
      consumes:
      - application/json
      description: create wallet
      operationId: create-wallet-admin
      parameters:
      - description: User ID
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.CreateWallet'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: uuid
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Wallet
      tags:
      - ADMIN
  /api/v1/admin/wallets/{wallet_uuid}:
    get:
      consumes:
      - application/json
      description: get wallet by uuid
      operationId: get-wallet-by-uuid-admin
      parameters:
      - description: Wallet uuid
        in: path
        name: wallet_uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Wallet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Wallet By UUID
      tags:
      - ADMIN
  /api/v1/wallet:
    get:
      consumes:
      - application/json
      description: get all wallets
      operationId: get-all-wallets
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllWalletsResponce'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get All Wallets
      tags:
      - Wallets
    post:
      consumes:
      - application/json
      description: update wallet
      operationId: update-wallet
      parameters:
      - description: Wallet query
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.WalletUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Wallet
      tags:
      - Wallets
  /api/v1/wallet/{wallet_uuid}:
    delete:
      consumes:
      - application/json
      description: delete wallet
      operationId: delete-wallet
      parameters:
      - description: Wallet UUID
        in: path
        name: wallet_uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Wallet
      tags:
      - Wallets
  /api/v1/wallet/new:
    post:
      consumes:
      - application/json
      description: create wallet
      operationId: create-wallet
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: uuid
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Wallet
      tags:
      - Wallets
  /api/v1/wallets/{wallet_uuid}:
    get:
      consumes:
      - application/json
      description: get wallet balance
      operationId: get-wallet-balance
      parameters:
      - description: Wallet UUID
        in: path
        name: wallet_uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllWalletsResponce'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Wallet Balance
      tags:
      - Wallets
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: login
      operationId: login
      parameters:
      - description: credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: SignIn
      tags:
      - Auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: create account
      operationId: create-account
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: SignUp
      tags:
      - Auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"